requirements.txt
fastapi>=0.103.0
uvicorn>=0.23.0
openai>=1.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
httpx>=0.24.0